LPS-186580 Tensorflow: move display context to its own package
