+++ modules/apps/document-library/document-library-asset-auto-tagger-tensorflow/src/main/java/com/liferay/document/library/asset/auto/tagger/tensorflow/internal/servlet/taglib/EditConfigurationJSPDynamicInclude.java
@@ -14,8 +14,8 @@
 
 package com.liferay.document.library.asset.auto.tagger.tensorflow.internal.servlet.taglib;
 
-import com.liferay.document.library.asset.auto.tagger.tensorflow.internal.EditConfigurationDisplayContext;
 import com.liferay.document.library.asset.auto.tagger.tensorflow.internal.configuration.TensorFlowImageAssetAutoTagProviderCompanyConfiguration;
+import com.liferay.document.library.asset.auto.tagger.tensorflow.internal.display.context.EditConfigurationDisplayContext;
 import com.liferay.document.library.asset.auto.tagger.tensorflow.internal.util.TensorFlowDownloadHelper;
 import com.liferay.petra.string.StringBundler;
 import com.liferay.portal.kernel.log.Log;
